+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 - added try...finally for cnx.close() and cursor.close() in myems-api
 - fixed overwritten property issues in myems-web
+- fixed overwritten property issues in myems-admin
 ### Removed
 
 ## [v6.4.0] - 2026-04-26